backend/requirements/base.txt
Django==5.2.12
djangorestframework==3.16.0
django-htmx==1.26.0
psycopg[binary]==3.2.10
whitenoise==6.9.0
gunicorn==23.0.0
openai==2.30.0
